Refactor: RentApiModel 클래스 이름 수정
RentAPIResponse 로 수정
